+++ cmd/helm/create.go
@@ -24,8 +24,8 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"k8s.io/helm/cmd/helm/helmpath"
 	"k8s.io/helm/pkg/chartutil"
+	"k8s.io/helm/pkg/helmpath"
 	"k8s.io/helm/pkg/proto/hapi/chart"
 )
