+++ cmd/helm/dependency_build.go
@@ -20,8 +20,8 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"k8s.io/helm/cmd/helm/helmpath"
 	"k8s.io/helm/pkg/downloader"
+	"k8s.io/helm/pkg/helmpath"
 )
 
 const dependencyBuildDesc = `
