+++ cmd/helm/home.go
@@ -21,7 +21,7 @@ import (
 	"io"
 
 	"github.com/spf13/cobra"
-	"k8s.io/helm/cmd/helm/helmpath"
+	"k8s.io/helm/pkg/helmpath"
 )
 
 var longHomeHelp = `
