+++ cmd/helm/repo_add_test.go
@@ -21,7 +21,7 @@ import (
 	"os"
 	"testing"
 
-	"k8s.io/helm/cmd/helm/helmpath"
+	"k8s.io/helm/pkg/helmpath"
 	"k8s.io/helm/pkg/repo"
 	"k8s.io/helm/pkg/repo/repotest"
 )
