+++ pkg/downloader/chart_downloader.go
@@ -27,7 +27,7 @@ import (
 	"path/filepath"
 	"strings"
 
-	"k8s.io/helm/cmd/helm/helmpath"
+	"k8s.io/helm/pkg/helmpath"
 	"k8s.io/helm/pkg/provenance"
 	"k8s.io/helm/pkg/repo"
 	"k8s.io/helm/pkg/urlutil"
