+++ pkg/helmpath/helmhome_windows_test.go
