+++ pkg/repo/repotest/server.go
@@ -24,7 +24,7 @@ import (
 
 	"github.com/ghodss/yaml"
 
-	"k8s.io/helm/cmd/helm/helmpath"
+	"k8s.io/helm/pkg/helmpath"
 	"k8s.io/helm/pkg/repo"
 )
